Move new userHasModuleAccess selector to modules store.
