--- assets/js/hooks/useModuleAcess.js
@@ -1,75 +1,0 @@
-/**
- * `useModuleAccess` hook.
- *
- * Site Kit by Google, Copyright 2022 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * Internal dependencies
- */
-import Data from 'googlesitekit-data';
-import { CORE_USER } from '../googlesitekit/datastore/user/constants';
-import { CORE_MODULES } from '../googlesitekit/modules/datastore/constants';
-const { useSelect } = Data;
-
-/**
- * Returns if the current logged in user has access to a given module.
- *
- * @since n.e.x.t
- *
- * @param {string} moduleSlug The module slug.
- * @return {Object} The result of the module access check and if the check is still loading.
- */
-function useModuleAccess( moduleSlug ) {
-	const storeName = useSelect( ( select ) =>
-		select( CORE_MODULES ).getModuleStoreName( moduleSlug )
-	);
-
-	const loggedInUserID = useSelect( ( select ) =>
-		select( CORE_USER ).getID()
-	);
-	const hasResolvedUser = useSelect( ( select ) =>
-		select( CORE_USER ).hasFinishedResolution( 'getUser' )
-	);
-
-	const hasModuleAccess = useSelect( ( select ) => {
-		const moduleOwnerID = select( storeName )?.getOwnerID();
-
-		if ( moduleOwnerID === loggedInUserID ) {
-			return true;
-		}
-
-		return select( CORE_MODULES ).hasModuleAccess( moduleSlug );
-	} );
-	const isLoadingModuleAccess = useSelect( ( select ) => {
-		const hasResolvedModuleOwner =
-			select( storeName )?.hasFinishedResolution( 'getSettings' );
-
-		const isResolvingModuleAccess = select( CORE_MODULES ).isResolving(
-			'hasModuleAccess',
-			[ moduleSlug ]
-		);
-
-		return (
-			! hasResolvedModuleOwner ||
-			! hasResolvedUser ||
-			isResolvingModuleAccess
-		);
-	} );
-
-	return { hasModuleAccess, isLoadingModuleAccess };
-}
-
-export default useModuleAccess;
